Refactored more reducers into their own modules.
